cut out validation to models/concerns
