+++ app/models/administrator.rb
@@ -1,13 +1,4 @@
 class Administrator < ApplicationRecord
-  before_validation do
-    self.email_for_index = email.downcase if email
-  end
-
-  def password=(raw_password)
-    if raw_password.is_a?(String)
-      self.hashed_password = BCrypt::Password.create(raw_password)
-    elsif raw_password.nil?
-      self.hashed_password = nil
-    end
-  end
+  include EmailHolder
+  include PasswordHolder
 end
